requirements.txt
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.9.0 ; python_version >= "3.13"
boto3==1.38.46 ; python_version >= "3.13"
botocore==1.38.46 ; python_version >= "3.13"
certifi==2025.6.15 ; python_version >= "3.13"
click==8.2.1 ; python_version >= "3.13"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.13"
dnspython==2.7.0 ; python_version >= "3.13"
email-validator==2.2.0 ; python_version >= "3.13"
fastapi-cli==0.0.7 ; python_version >= "3.13"
fastapi==0.115.13 ; python_version >= "3.13"
h11==0.16.0 ; python_version >= "3.13"
httpcore==1.0.9 ; python_version >= "3.13"
httptools==0.6.4 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
jmespath==1.0.1 ; python_version >= "3.13"
mangum==0.19.0 ; python_version >= "3.13"
markdown-it-py==3.0.0 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
mdurl==0.1.2 ; python_version >= "3.13"
pydantic-core==2.33.2 ; python_version >= "3.13" ; platform==
pydantic==2.11.7 ; python_version >= "3.13"
pygments==2.19.2 ; python_version >= "3.13"
pyjwt==2.10.1 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
python-dotenv==1.1.1 ; python_version >= "3.13"
python-multipart==0.0.20 ; python_version >= "3.13"
pyyaml==6.0.2 ; python_version >= "3.13"
rich-toolkit==0.14.7 ; python_version >= "3.13"
rich==14.0.0 ; python_version >= "3.13"
s3transfer==0.13.0 ; python_version >= "3.13"
shellingham==1.5.4 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
starlette==0.46.2 ; python_version >= "3.13"
structlog==25.4.0 ; python_version >= "3.13"
tenacity==9.1.2 ; python_version >= "3.13"
typer==0.16.0 ; python_version >= "3.13"
typing-extensions==4.14.0 ; python_version >= "3.13"
typing-inspection==0.4.1 ; python_version >= "3.13"
urllib3==2.5.0 ; python_version >= "3.13"
uvicorn==0.34.3 ; python_version >= "3.13"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.13"
watchfiles==1.1.0 ; python_version >= "3.13"
websockets==15.0.1 ; python_version >= "3.13"
